+++ inst/shiny/global.R
@@ -1,0 +1,7 @@
+library(shiny)
+library(brickset)
+library(dplyr)
+
+data(legosets, package = 'brickset')
+
+default_view_cols <- c('setID','name','year','theme','released','pieces','minifigs','availability','US_retailPrice')