+++ NEWS.md
@@ -1,9 +1,10 @@
 # brickset (development version)
 
 * Added number, thumbnailURL, and imageURL to the data.frame returned by the getSets function.
-* Added check for errors in returned JSON. For example, exceeding the API limit will return a user friendly message.
+* Will check for errors in returned JSON. For example, exceeding the API limit will return a user friendly message.
 * Added code of conduct.
 * Fixed location of README figures.
+* Added reviews data frame through Github releases. 
 
 # brickset 2022.0.0